Rename co.rsk.peg package to co.rsk.bridge
