+++ src/main/java/co/rsk/federate/BtcToRskClient.java
@@ -10,14 +10,14 @@ import co.rsk.federate.io.*;
 import co.rsk.federate.timing.TurnScheduler;
 import co.rsk.net.NodeBlockProcessor;
 import co.rsk.panic.PanicProcessor;
-import co.rsk.peg.BridgeUtils;
-import co.rsk.peg.Federation;
-import co.rsk.peg.PeginInformation;
-import co.rsk.peg.btcLockSender.BtcLockSender.TxSenderAddressType;
-import co.rsk.peg.pegininstructions.PeginInstructionsException;
-import co.rsk.peg.pegininstructions.PeginInstructionsProvider;
+import co.rsk.bridge.BridgeUtils;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.PeginInformation;
+import co.rsk.bridge.btcLockSender.BtcLockSender.TxSenderAddressType;
+import co.rsk.bridge.pegininstructions.PeginInstructionsException;
+import co.rsk.bridge.pegininstructions.PeginInstructionsProvider;
 import com.google.common.annotations.VisibleForTesting;
-import co.rsk.peg.btcLockSender.BtcLockSenderProvider;
+import co.rsk.bridge.btcLockSender.BtcLockSenderProvider;
 import com.google.common.collect.Lists;
 import org.bitcoinj.core.*;
 import org.bitcoinj.store.BlockStoreException;
