+++ src/main/java/co/rsk/federate/FedNodeRunner.java
@@ -47,10 +47,10 @@ import co.rsk.federate.signing.hsm.message.ReleaseCreationInformationGetter;
 import co.rsk.federate.signing.hsm.message.SignerMessageBuilderFactory;
 import co.rsk.federate.signing.hsm.requirements.AncestorBlockUpdater;
 import co.rsk.federate.signing.hsm.requirements.ReleaseRequirementsEnforcer;
-import co.rsk.peg.Federation;
-import co.rsk.peg.FederationMember;
-import co.rsk.peg.btcLockSender.BtcLockSenderProvider;
-import co.rsk.peg.pegininstructions.PeginInstructionsProvider;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.FederationMember;
+import co.rsk.bridge.btcLockSender.BtcLockSenderProvider;
+import co.rsk.bridge.pegininstructions.PeginInstructionsProvider;
 import org.bitcoinj.core.Context;
 import org.bouncycastle.util.encoders.Hex;
 import org.ethereum.crypto.ECKey;
