+++ src/main/java/co/rsk/federate/FederatorSupport.java
@@ -8,16 +8,14 @@ import co.rsk.federate.adapter.ThinConverter;
 import co.rsk.federate.bitcoin.BitcoinPeerFactory;
 import co.rsk.federate.config.FedNodeSystemProperties;
 import co.rsk.federate.signing.ECDSASigner;
-import co.rsk.peg.Bridge;
-import co.rsk.peg.BridgeMethods;
-import co.rsk.peg.FederationMember;
-import co.rsk.peg.StateForFederator;
+import co.rsk.bridge.Bridge;
+import co.rsk.bridge.FederationMember;
+import co.rsk.bridge.StateForFederator;
 import org.bitcoinj.core.PartialMerkleTree;
 import org.bitcoinj.core.PeerAddress;
 import org.bitcoinj.core.Sha256Hash;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.core.Blockchain;
-import org.ethereum.core.CallTransaction;
 import org.ethereum.crypto.ECKey;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
