+++ src/main/java/co/rsk/federate/bitcoin/BitcoinWrapperImpl.java
@@ -4,12 +4,12 @@ import co.rsk.bitcoinj.core.BtcTransaction;
 import co.rsk.config.BridgeConstants;
 import co.rsk.federate.FederatorSupport;
 import co.rsk.federate.adapter.ThinConverter;
-import co.rsk.peg.BridgeUtils;
-import co.rsk.peg.Federation;
-import co.rsk.peg.PeginInformation;
-import co.rsk.peg.btcLockSender.BtcLockSenderProvider;
-import co.rsk.peg.pegininstructions.PeginInstructionsException;
-import co.rsk.peg.pegininstructions.PeginInstructionsProvider;
+import co.rsk.bridge.BridgeUtils;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.PeginInformation;
+import co.rsk.bridge.btcLockSender.BtcLockSenderProvider;
+import co.rsk.bridge.pegininstructions.PeginInstructionsException;
+import co.rsk.bridge.pegininstructions.PeginInstructionsProvider;
 import co.rsk.util.MaxSizeHashMap;
 import java.util.Collection;
 import java.util.Collections;
