+++ src/main/java/co/rsk/federate/btcreleaseclient/BtcReleaseClient.java
@@ -29,11 +29,11 @@ import co.rsk.federate.signing.hsm.requirements.ReleaseRequirementsEnforcer;
 import co.rsk.federate.signing.hsm.requirements.ReleaseRequirementsEnforcerException;
 import co.rsk.net.NodeBlockProcessor;
 import co.rsk.panic.PanicProcessor;
-import co.rsk.peg.Bridge;
-import co.rsk.peg.BridgeEvents;
-import co.rsk.peg.BridgeUtils;
-import co.rsk.peg.Federation;
-import co.rsk.peg.StateForFederator;
+import co.rsk.bridge.Bridge;
+import co.rsk.bridge.BridgeEvents;
+import co.rsk.bridge.BridgeUtils;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.StateForFederator;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
