+++ src/test/java/co/rsk/federate/BtcToRskClientBuilder.java
@@ -10,10 +10,10 @@ import co.rsk.federate.bitcoin.BitcoinWrapper;
 import co.rsk.federate.io.BtcToRskClientFileData;
 import co.rsk.federate.io.BtcToRskClientFileReadResult;
 import co.rsk.federate.io.BtcToRskClientFileStorage;
-import co.rsk.peg.Federation;
-import co.rsk.peg.btcLockSender.BtcLockSenderProvider;
-import co.rsk.peg.pegininstructions.PeginInstructionsException;
-import co.rsk.peg.pegininstructions.PeginInstructionsProvider;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.btcLockSender.BtcLockSenderProvider;
+import co.rsk.bridge.pegininstructions.PeginInstructionsException;
+import co.rsk.bridge.pegininstructions.PeginInstructionsProvider;
 import java.io.IOException;
 import java.util.Optional;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
