+++ src/test/java/co/rsk/federate/BtcToRskClientTest.java
@@ -19,14 +19,14 @@ import co.rsk.federate.mock.SimpleBlock;
 import co.rsk.federate.mock.SimpleBtcTransaction;
 import co.rsk.federate.mock.SimpleFederatorSupport;
 import co.rsk.net.NodeBlockProcessor;
-import co.rsk.peg.BridgeUtils;
-import co.rsk.peg.Federation;
-import co.rsk.peg.btcLockSender.BtcLockSender;
-import co.rsk.peg.btcLockSender.BtcLockSender.TxSenderAddressType;
-import co.rsk.peg.btcLockSender.BtcLockSenderProvider;
-import co.rsk.peg.btcLockSender.P2shP2wpkhBtcLockSender;
-import co.rsk.peg.pegininstructions.PeginInstructionsException;
-import co.rsk.peg.pegininstructions.PeginInstructionsProvider;
+import co.rsk.bridge.BridgeUtils;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.btcLockSender.BtcLockSender;
+import co.rsk.bridge.btcLockSender.BtcLockSender.TxSenderAddressType;
+import co.rsk.bridge.btcLockSender.BtcLockSenderProvider;
+import co.rsk.bridge.btcLockSender.P2shP2wpkhBtcLockSender;
+import co.rsk.bridge.pegininstructions.PeginInstructionsException;
+import co.rsk.bridge.pegininstructions.PeginInstructionsProvider;
 import com.google.common.collect.Lists;
 import org.bitcoinj.core.*;
 import org.bitcoinj.script.Script;
