+++ src/test/java/co/rsk/federate/FederationProviderFromFederatorSupportTest.java
@@ -6,10 +6,10 @@ import co.rsk.bitcoinj.core.NetworkParameters;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.config.BridgeConstants;
 import co.rsk.config.BridgeTestNetConstants;
-import co.rsk.peg.ErpFederation;
-import co.rsk.peg.Federation;
-import co.rsk.peg.FederationMember;
-import co.rsk.peg.P2shErpFederation;
+import co.rsk.bridge.ErpFederation;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.FederationMember;
+import co.rsk.bridge.P2shErpFederation;
 import org.bouncycastle.util.encoders.Hex;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.crypto.ECKey;
