+++ src/test/java/co/rsk/federate/FederationWatcherTest.java
@@ -3,8 +3,8 @@ package co.rsk.federate;
 import co.rsk.bitcoinj.core.BtcECKey;
 import co.rsk.bitcoinj.core.NetworkParameters;
 import co.rsk.federate.signing.utils.TestUtils;
-import co.rsk.peg.Federation;
-import co.rsk.peg.FederationMember;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.FederationMember;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.facade.Ethereum;
 import org.ethereum.listener.EthereumListenerAdapter;
