+++ src/test/java/co/rsk/federate/FederatorSupportTest.java
@@ -3,8 +3,8 @@ package co.rsk.federate;
 import co.rsk.config.BridgeRegTestConstants;
 import co.rsk.federate.adapter.ThinConverter;
 import co.rsk.federate.config.TestSystemProperties;
-import co.rsk.peg.Bridge;
-import co.rsk.peg.BridgeMethods;
+import co.rsk.bridge.Bridge;
+import co.rsk.bridge.BridgeMethods;
 import org.bitcoinj.core.*;
 import org.bouncycastle.util.encoders.Hex;
 import org.ethereum.core.Blockchain;
