+++ src/test/java/co/rsk/federate/bitcoin/BitcoinWrapperImplTest.java
@@ -15,11 +15,11 @@ import co.rsk.config.BridgeConstants;
 import co.rsk.config.BridgeRegTestConstants;
 import co.rsk.federate.FederatorSupport;
 import co.rsk.federate.adapter.ThinConverter;
-import co.rsk.peg.Federation;
-import co.rsk.peg.btcLockSender.BtcLockSender;
-import co.rsk.peg.btcLockSender.BtcLockSenderProvider;
-import co.rsk.peg.pegininstructions.PeginInstructionsException;
-import co.rsk.peg.pegininstructions.PeginInstructionsProvider;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.btcLockSender.BtcLockSender;
+import co.rsk.bridge.btcLockSender.BtcLockSenderProvider;
+import co.rsk.bridge.pegininstructions.PeginInstructionsException;
+import co.rsk.bridge.pegininstructions.PeginInstructionsProvider;
 import java.io.File;
 import java.util.Collections;
 import java.util.List;
