+++ src/test/java/co/rsk/federate/bitcoin/KitTest.java
@@ -5,8 +5,8 @@ import co.rsk.config.BridgeRegTestConstants;
 import co.rsk.federate.FederatorSupport;
 import co.rsk.federate.adapter.ThinConverter;
 import co.rsk.federate.signing.utils.TestUtils;
-import co.rsk.peg.btcLockSender.BtcLockSenderProvider;
-import co.rsk.peg.pegininstructions.PeginInstructionsProvider;
+import co.rsk.bridge.btcLockSender.BtcLockSenderProvider;
+import co.rsk.bridge.pegininstructions.PeginInstructionsProvider;
 import org.bitcoinj.core.*;
 import org.bitcoinj.wallet.Wallet;
 import org.junit.BeforeClass;
