+++ src/test/java/co/rsk/federate/btcreleaseclient/BtcReleaseClientTest.java
@@ -53,10 +53,10 @@ import co.rsk.federate.signing.hsm.requirements.ReleaseRequirementsEnforcer;
 import co.rsk.federate.signing.hsm.requirements.ReleaseRequirementsEnforcerException;
 import co.rsk.federate.signing.utils.TestUtils;
 import co.rsk.net.NodeBlockProcessor;
-import co.rsk.peg.ErpFederation;
-import co.rsk.peg.Federation;
-import co.rsk.peg.FederationMember;
-import co.rsk.peg.StateForFederator;
+import co.rsk.bridge.ErpFederation;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.FederationMember;
+import co.rsk.bridge.StateForFederator;
 import java.math.BigInteger;
 import java.time.Instant;
 import java.util.*;
