+++ src/test/java/co/rsk/federate/signing/utils/TestUtils.java
@@ -5,8 +5,8 @@ import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.wallet.RedeemData;
 import co.rsk.core.BlockDifficulty;
 import co.rsk.crypto.Keccak256;
-import co.rsk.peg.Federation;
-import co.rsk.peg.FederationMember;
+import co.rsk.bridge.Federation;
+import co.rsk.bridge.FederationMember;
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.ethereum.core.Block;
 import org.ethereum.core.BlockHeader;
